New module with integration tests common logic #173
