+++ integration/util/config.go
@@ -12,14 +12,6 @@
 
 package util
 
-import (
-	"encoding/json"
-	"fmt"
-	"time"
-
-	MQTT "github.com/eclipse/paho.mqtt.golang"
-)
-
 // TestConfiguration is common IT configuration
 type TestConfiguration struct {
 	LocalBroker              string `env:"LOCAL_BROKER" envDefault:"tcp://localhost:1883"`
@@ -32,43 +24,4 @@ type TestConfiguration struct {
 	DigitalTwinAPIPassword string `env:"DIGITAL_TWIN_API_PASSWORD" envDefault:"ditto"`
 
 	WsEventTimeoutMs int `env:"WS_EVENT_TIMEOUT_MS" envDefault:"30000"`
-}
-
-// ThingConfiguration is thing configuration info for Edge
-type ThingConfiguration struct {
-	DeviceID string `json:"deviceId"`
-	TenantID string `json:"tenantId"`
-	PolicyID string `json:"policyId"`
-}
-
-// GetThingConfiguration retrieves ThingConfig using specified client
-func GetThingConfiguration(mqttClient MQTT.Client) (*ThingConfiguration, error) {
-	type result struct {
-		cfg *ThingConfiguration
-		err error
-	}
-
-	ch := make(chan result)
-
-	if token := mqttClient.Subscribe("edge/thing/response", 1, func(client MQTT.Client, message MQTT.Message) {
-		var cfg ThingConfiguration
-		if err := json.Unmarshal(message.Payload(), &cfg); err != nil {
-			ch <- result{nil, err}
-		}
-		ch <- result{&cfg, nil}
-	}); token.Wait() && token.Error() != nil {
-		return nil, token.Error()
-	}
-
-	if token := mqttClient.Publish("edge/thing/request", 1, false, ""); token.Wait() && token.Error() != nil {
-		return nil, token.Error()
-	}
-
-	timeout := 5 * time.Second
-	select {
-	case result := <-ch:
-		return result.cfg, result.err
-	case <-time.After(timeout):
-		return nil, fmt.Errorf("thing config not received in %v", timeout)
-	}
 }
