+++ integration/util/mqtt_client.go
@@ -15,6 +15,7 @@ package util
 import (
 	"encoding/json"
 	"errors"
+	"fmt"
 	"time"
 
 	"github.com/google/uuid"
@@ -52,4 +53,45 @@ func SendMQTTMessage(cfg *TestConfiguration, client MQTT.Client, topic string, m
 		return errors.New("timeout")
 	}
 	return token.Error()
+}
+
+// ThingConfiguration is thing configuration info for Edge
+type ThingConfiguration struct {
+	DeviceID string `json:"deviceId"`
+	TenantID string `json:"tenantId"`
+	PolicyID string `json:"policyId"`
+}
+
+// GetThingConfiguration retrieves ThingConfig using specified client
+func GetThingConfiguration(mqttClient MQTT.Client) (*ThingConfiguration, error) {
+	type result struct {
+		cfg *ThingConfiguration
+		err error
+	}
+
+	ch := make(chan result)
+
+	if token := mqttClient.Subscribe("edge/thing/response", 1, func(client MQTT.Client, message MQTT.Message) {
+		var cfg ThingConfiguration
+		if err := json.Unmarshal(message.Payload(), &cfg); err != nil {
+			ch <- result{nil, err}
+		}
+		ch <- result{&cfg, nil}
+	}); token.Wait() && token.Error() != nil {
+		return nil, token.Error()
+	}
+
+	defer mqttClient.Unsubscribe("edge/thing/response")
+
+	if token := mqttClient.Publish("edge/thing/request", 1, false, ""); token.Wait() && token.Error() != nil {
+		return nil, token.Error()
+	}
+
+	timeout := 5 * time.Second
+	select {
+	case result := <-ch:
+		return result.cfg, result.err
+	case <-time.After(timeout):
+		return nil, fmt.Errorf("thing config not received in %v", timeout)
+	}
 }
